Decode signed ORC timestamp nanoseconds and truncate to the unit
The Apache ORC C++ writer stores the negative fraction that pyarrow gives a
pre-1970 timestamp as is, so 1969-07-20 20:17:40.5 arrives as -14182939
seconds and -500000000 nanoseconds. decode cast that to u64, and undoing
the trailing-zero encoding then overflowed: a debug build panicked and a
release build failed the whole file. The nanoseconds now decode as signed,
as the C++ reader does, and the arithmetic runs in i128, so no input can
overflow it.

decode_timestamp also refused any value with digits below the requested
unit, so one nanosecond value failed a file read at microsecond
precision. It now truncates toward zero, as Arrow's cast between
timestamp units does, and still fails a value that overflows i64.

alltypes_test and test_lzo fail before and after this change because the
bounded reader refuses LZO streams.

+++ src/error.rs
@@ -65,8 +65,8 @@ pub enum OrcError {
     DecodeTimestamp {
         #[snafu(implicit)]
         location: Location,
-        seconds: i64,
-        nanoseconds: u64,
+        seconds: i128,
+        nanoseconds: i128,
         to_time_unit: TimeUnit,
     },
